// src/types/openapi_MS.ts
export interface components {
    schemas: {
// ...
        ResLoginDTO: {
            user?: components["schemas"]["UserLogin"];
            role?: components["schemas"]["Role"];
            access_token?: string;
        };
// ...
        UserGetAccount: {
            user?: components["schemas"]["UserLogin"];
            role?: components["schemas"]["Role"];
            grades?: components["schemas"]["GradeAccess"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}